[add]weeklyにランキング実装
