+++ src/components/common/styles.ts
@@ -1,0 +1,5 @@
+export const MIDNIGHT_BLUE = "#003366";
+export const PACIFIC_BLUE = "#0099CC";
+export const ANAKIWA = "#99CCFF";
+export const BAHAMA_BLUE = "#006699";
+export const MALIBU = "#66CCFF";
