+++ src/components/layouts/ToggleGroup.jsx
@@ -2,7 +2,7 @@
 import ToggleButton from "@mui/material/ToggleButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import CheckIcon from "@mui/icons-material/Check";
-import { Text } from "../../components/common/Text";
+import { Text } from "../common/Text";
 
 export const ToggleGroup = (props) => {
   const { toggleHandler, timeUnit } = props;
